Use unsigned ints for location

+++ MikeRectangle.cpp
@@ -29,35 +29,35 @@ void MikeRectangle::Draw(LPDIRECT3DDEVICE9 device) {
 
 	VERTEX_2D_DIF verts[5];
 
-	int xOffset = (width * scale) / 2;
-	int yOffset = (height * scale) / 2;
+	unsigned int xOffset = (width * scale) / 2;
+	unsigned int yOffset = (height * scale) / 2;
 
-	verts[0].x = x - xOffset;
-	verts[0].y = y - yOffset;
+	verts[0].x = (float)(x - xOffset);
+	verts[0].y = (float)(y - yOffset);
 	verts[0].z = 1;
 	verts[0].rhw = 1;
 	verts[0].color = graphicsNS::MAGENTA;
 
-	verts[1].x = x + xOffset;
-	verts[1].y = y - yOffset;
+	verts[1].x = (float)(x + xOffset);
+	verts[1].y = (float)(y - yOffset);
 	verts[1].z = 1;
 	verts[1].rhw = 1;
 	verts[1].color = graphicsNS::MAGENTA;
 
-	verts[2].x = x + xOffset;
-	verts[2].y = y + yOffset;
+	verts[2].x = (float)(x + xOffset);
+	verts[2].y = (float)(y + yOffset);
 	verts[2].z = 1;
 	verts[2].rhw = 1;
 	verts[2].color = graphicsNS::MAGENTA;
 
-	verts[3].x = x - xOffset;
-	verts[3].y = y + yOffset;
+	verts[3].x = (float)(x - xOffset);
+	verts[3].y = (float)(y + yOffset);
 	verts[3].z = 1;
 	verts[3].rhw = 1;
 	verts[3].color = graphicsNS::MAGENTA;
 
-	verts[4].x = x - xOffset;
-	verts[4].y = y - yOffset;
+	verts[4].x = (float)(x - xOffset);
+	verts[4].y = (float)(y - yOffset);
 	verts[4].z = 1;
 	verts[4].rhw = 1;
 	verts[4].color = graphicsNS::MAGENTA;
